perf: separate prebuilt startup and defer unused SDK loading
