+++ package.json
@@ -28,7 +28,9 @@
     "test:live:vision-chat": "node --env-file-if-exists=.env ./node_modules/vitest/vitest.mjs run packages/testkit/src/openai-compatible-model-driver.live.test.ts -t 'Vision Chat eager image' --reporter=verbose",
     "test:live:vision-responses": "node --env-file-if-exists=.env ./node_modules/vitest/vitest.mjs run packages/testkit/src/openai-compatible-model-driver.live.test.ts -t 'Vision Responses lazy image' --reporter=verbose",
     "test:live:skills": "node --env-file-if-exists=.env ./node_modules/vitest/vitest.mjs run packages/testkit/src/agent-skills.live.test.ts --reporter=verbose",
-    "typecheck": "tsc -b --pretty false"
+    "typecheck": "tsc -b --pretty false",
+    "adam:run": "node apps/cli/dist/main.js",
+    "tui:run": "node apps/tui/dist/main.js"
   },
   "devDependencies": {
     "@biomejs/biome": "2.5.8",
